// src/Mithrill.MonsterBook.WebApi/monsterbook/src/app/npc-dashboard/models/npc.model.ts
import { Difficulty } from '../../core/model/difficulty.model';
import { Race } from '../../core/model/race.model';
import { Armor } from './armor.model';
import { Flaw } from './flaw.model';
import { Merit } from './merit.model';
import { Skill } from './skill.model';
import { Weapon } from './weapon.model';

export interface Npc {
  id: number;
  name: string;
  strengthMin: number;
  strengthMax: number;
  vitalityMin: number;
  vitalityMax: number;
  bodyMin: number;
  bodyMax: number;
  agilityMin: number;
  agilityMax: number;
  dexterityMin: number;
  dexterityMax: number;
  intelligenceMin: number;
  intelligenceMax: number;
  willpowerMin: number;
  willpowerMax: number;
  emotionMin: number;
  emotionMax: number;
  karmaMin: number;
  karmaMax: number;
  hitPointMin: number;
  hitPointMax: number;
  manaMin: number;
  manaMax: number;
  powerPointMin: number;
  powerPointMax: number;
  race: Race;
  difficulty: Difficulty;
  skills: Skill[];
  merits: Merit[];
  flaws: Flaw[];
  weapons: Weapon[];
  armors: Armor[];
}